Projekt refactored away from MVVM
