+++ WHUT/Business/TournamentViewModel.cs
@@ -3,9 +3,9 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using WHUT.Model;
+using WHUT.Domain;
 
-namespace WHUT.ViewModel
+namespace WHUT.Business
 {
     public class TournamentViewModel
     {
