+++ WHUT/Domain/Tournament.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace WHUT.Model
+namespace WHUT.Domain
 {
     public class Tournament
     {
