--- WHUT/View/Content/NewTournament.xaml
@@ -1,14 +1,0 @@
-﻿<Page x:Class="WHUT.View.Content.NewTournament"
-      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-      xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-      xmlns:local="clr-namespace:WHUT.View.Content"
-      mc:Ignorable="d" 
-      d:DesignHeight="450" d:DesignWidth="800"
-      Title="NewTournament">
-
-    <Grid>
-        
-    </Grid>
-</Page>
